pyproject.toml
[project]
name = "ocr-pipeline"
version = "0.1.0"
description = "Offline OCR pipeline for large mind map images using PaddleOCR."
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "networkx>=3.4",
    "numpy>=2.2",
    "opencv-python>=4.11",
    "pillow>=11.1",
    "pydantic>=2.11",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6",
    "shapely>=2.0",
    "typer>=0.16",
]

[project.optional-dependencies]
ocr = [
    "paddleocr>=2.8.1",
]

[project.scripts]
ocr-pipeline = "ocr_pipeline.cli:app"

[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ocr_pipeline"]
include = ["src/ocr_pipeline/editor_template.html"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--basetemp=tests_tmp"]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "ruff>=0.11",
]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["B008", "B905", "E501"]
